feat/#27 - 채팅방시작구현

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/chat/dto/ChatRequestDto.java
@@ -7,9 +7,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ChatRequestDto {
 
-    private Long roomId;
+    private Integer roomId;
 
-    private Long senderId;
+    private Integer senderId;
 
     private String message;
 }
